HttpClient tip unrelated to Http/WebHook binding
The HttpClient instantiation tip is valuable whenever you are making outbound http calls, not just when you have an Http or WebHook binding. Http and WebHook bindings are related to Http/WebHook triggers. Regardless of the function trigger, the function could make outbound http calls using the HttpClient, which should be instantiated properly, as mentioned in the related article.